clash: fix 6 errors + 11 warnings from Revit 2025+ Windows build
ERRORS (6):

ClashSession.cs + ClashExportContext.cs — ExporterIFCUtils (CS0103):
ExporterIFCUtils moved to RevitAPIIFC.dll in Revit 2025+ which
StingTools doesn't reference. Dropped `using Autodesk.Revit.DB.IFC;`
from both files and replaced CreateSubElementGUID() with UniqueId
fallback. BCF export and ClashIdentity.Compute already treat IfcGuid
as optional — UniqueId gives a stable per-element identifier that's
always available without the IFC assembly.

ClashExportContext.cs — LinkNode.GetSymbolId() (CS1061):
Method removed in Revit 2025. Replaced with guid.GetHashCode() as a
stable synthetic link instance id. ClashElementKey.Equals composes
DocGuid + LinkInstanceElementId + ElementId so two elements with the
same internal id in different links never collide.

ClashRuleLibrary.cs — Exception not found (CS0246):
H9's catch (Exception ex) needed `using System;`. Added.

BIMCoordinationCenter.cs — _doc not found (CS0103):
BCC class has no _doc field — code throughout uses
`StingCommandHandler.CurrentApp?.ActiveUIDocument?.Document`. My
rec-4 BuildClashTab used _doc assuming convention; fixed to follow
the established pattern with null-guards.

WARNINGS (11):

CS0618 ElementId.IntegerValue (9 sites): Revit 2024+ replaced with
Value (returns long). Cast to int at ClashElementKey boundary —
Revit user-authored element ids fit well within int range. Files:
LiveClashUpdater.cs, LiveClashFlag.cs, ClashSession.cs,
ClashExportContext.cs, CategoryHelper.cs.

CS0618 ElementId(int) ctor: use ElementId((long)…) overload in
ClashSession.cs, LiveClashFlag.cs, ClashRunCommand.cs.

CS0169 BcfSnapshotter._tempView never used: removed (was a
never-assigned cache field from the prompt draft; temp view is a
local variable inside RenderSnapshot now).

CS0067 ClashRowViewModel.PropertyChanged never raised: pragma
suppressed with a comment explaining it's forward-compat for
inline editing (DataGrid today rebinds on Rows.Clear/Add).

35/35 local tests still pass. Windows build should restore + compile
cleanly — user's Linux sandbox can't run dotnet build against
net8.0-windows + Revit refs so confirmation goes through their
Windows dev machine.

+++ StingTools/Clash/ClashRunCommand.cs
@@ -230,7 +230,8 @@ namespace StingTools.Core.Clash
                     // Resolve the element from its owning document (host OR link).
                     if (docByGuid.TryGetValue(key.DocGuid ?? "", out var resolvedDoc))
                         owningDoc = resolvedDoc;
-                    el = owningDoc?.GetElement(new ElementId(key.ElementId));
+                    // Revit 2024+: ElementId(int) ctor obsolete.
+                    el = owningDoc?.GetElement(new ElementId((long)key.ElementId));
                 }
                 catch (Exception ex) { StingLog.Warn($"BuildFactsByKey GetElement({key}): {ex.Message}"); }
